pyproject.toml
[project]
name = "local-llm"
version = "0.1.0"
description = "An experimental project to have a local research assistant and RAG model working on my codebases."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.5.20",
    "datasets>=3.2.0",
    "einops>=0.8.0",
    "fastapi>=0.115.7",
    "langchain>=0.3.15",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.15",
    "langchain-ollama>=0.2.2",
    "langgraph>=0.2.67",
    "libmagic>=1.0",
    "nltk>=3.9.1",
    "ollama>=0.4.7",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=3.4.0",
    "tavily-python>=0.5.0",
    "transformers>=4.48.1",
    "unstructured>=0.16.15",
    "uvicorn>=0.34.0",
    "websockets>=14.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
